+++ src/services/mute.ts
@@ -1,28 +1,22 @@
 import { Service, PlatformAccessory, CharacteristicValue, CharacteristicGetCallback, Characteristic } from 'homebridge';
 import { SonosPlatform } from '../platform';
-import { PlatformDeviceManager } from '../platformDeviceManager';
+import { SonosDeviceManager } from '../helpers/sonosDeviceManager';
+import { ServiceNames } from '../models/enums';
 
 export class MuteService {
     private service: Service | undefined;
-    private readonly device: PlatformDeviceManager;
-    private name: string = 'Mute';
+    private readonly device: SonosDeviceManager;
+    private name: string = ServiceNames.MuteService;
 
     constructor(
         private readonly platform: SonosPlatform,
         private readonly accessory: PlatformAccessory,
-        sonosDevice: PlatformDeviceManager,
+        sonosDevice: SonosDeviceManager,
         displayOrder: number
     ) {
         this.device = sonosDevice;
 
-        //Optional Mute Switch, so remove it incase the user has chosen no mute.
-        let oldMuteService = this.accessory.getService(this.name);
-        oldMuteService ? this.accessory.removeService(oldMuteService) : null;
-
-        if (!this.platform.config.muteSwitch) return;
-
-        this.service = this.accessory.addService(this.platform.Service.Switch, this.name, this.name);
-
+        this.service = this.accessory.getService(this.name) || this.accessory.addService(this.platform.Service.Switch, this.name, this.name);
         this.service.addOptionalCharacteristic(this.platform.Characteristic.ConfiguredName);
         this.service.getCharacteristic(this.platform.Characteristic.ConfiguredName).setValue(this.name);