+++ src/services/volumeControls.ts
@@ -1,42 +1,33 @@
 import { Service, PlatformAccessory, CharacteristicValue } from 'homebridge';
 import { RotationSpeed, Brightness } from 'hap-nodejs/dist/lib/definitions/CharacteristicDefinitions';
-import { Fan, Lightbulb } from 'hap-nodejs/dist/lib/definitions/ServiceDefinitions';
-import { VolumeOptions } from '../constants';
 import { SonosPlatform } from '../platform';
-import { PlatformDeviceManager } from '../platformDeviceManager';
+import { SonosDeviceManager } from '../helpers/sonosDeviceManager';
+import { ServiceNames, VolumeOptions, DeviceEvents } from '../models/enums';
 
 export class VolumeControlService {
-    private readonly device: PlatformDeviceManager;
-    private service: Service | undefined;
-    private name: string = 'Volume';
-    private serviceType: typeof Fan | typeof Lightbulb | undefined;
-    private volumeCharacteristic: typeof RotationSpeed | typeof Brightness | undefined;
+    private readonly device: SonosDeviceManager;
+    private service!: Service;
+    private name: string = ServiceNames.VolumeService;
+    private volumeCharacteristic!: typeof RotationSpeed | typeof Brightness;
 
     constructor(
         private readonly platform: SonosPlatform,
         private readonly accessory: PlatformAccessory,
-        sonosDevice: PlatformDeviceManager,
+        sonosDevice: SonosDeviceManager,
         displayOrder: number
     ) {
         this.device = sonosDevice;
 
+        //Volume controls could change so check to see if that they haven't
         let currentVolumeService = this.accessory.getService(this.name);
-        currentVolumeService ? this.accessory.removeService(currentVolumeService) : null;
-
         switch (this.platform.config.volume) {
             case VolumeOptions.Fan:
-                this.serviceType = this.platform.Service.Fan;
-                this.volumeCharacteristic = this.platform.Characteristic.RotationSpeed;
+                this.checkAndSetupFan(currentVolumeService);
                 break;
             case VolumeOptions.Lightbulb:
-                this.serviceType = this.platform.Service.Lightbulb;
-                this.volumeCharacteristic = this.platform.Characteristic.Brightness;
+                this.checkAndSetupBulb(currentVolumeService);
                 break;
-            case VolumeOptions.None:
-                return;
         }
-
-        this.service = this.accessory.addService(this.serviceType!, this.name, this.name);
 
         this.service.addOptionalCharacteristic(this.platform.Characteristic.ConfiguredName);
         this.service.getCharacteristic(this.platform.Characteristic.ConfiguredName).setValue(this.name);
@@ -47,6 +38,32 @@ export class VolumeControlService {
         this.service.getCharacteristic(this.platform.Characteristic.On).onGet(this.handleMuteGet.bind(this)).onSet(this.handleMuteSet.bind(this));
 
         this.service.getCharacteristic(this.volumeCharacteristic!).onSet(this.handleVolumeSet.bind(this));
+
+        this.device.on(DeviceEvents.DeviceVolumeUpdate, (volume: number) => {
+            this.updateCharacteristic(volume);
+        });
+    }
+
+    private checkAndSetupFan(currentVolumeService?: Service) {
+        this.volumeCharacteristic = this.platform.Characteristic.RotationSpeed;
+
+        if (currentVolumeService && currentVolumeService.testCharacteristic(this.platform.Characteristic.RotationSpeed)) {
+            this.service = currentVolumeService;
+            return;
+        }
+        if (currentVolumeService) this.accessory.removeService(currentVolumeService);
+        this.service = this.accessory.addService(this.platform.Service.Fan, this.name, this.name);
+    }
+
+    private checkAndSetupBulb(currentVolumeService?: Service) {
+        this.volumeCharacteristic = this.platform.Characteristic.Brightness;
+
+        if (currentVolumeService && currentVolumeService.testCharacteristic(this.platform.Characteristic.Brightness)) {
+            this.service = currentVolumeService;
+            return;
+        }
+        if (currentVolumeService) this.accessory.removeService(currentVolumeService);
+        this.service = this.accessory.addService(this.platform.Service.Lightbulb, this.name, this.name);
     }
 
     private async handleMuteGet() {
@@ -62,11 +79,7 @@ export class VolumeControlService {
         this.device.setVolume(value as number);
     }
 
-    public updateCharacteristic(volume: number) {
-        this.platform.log.debug('Update Volume Triggered');
-        if (this.service) {
-            this.platform.log.debug(`Setting Fan To: ${volume}`);
-            this.service.updateCharacteristic(this.volumeCharacteristic!, volume);
-        }
+    private updateCharacteristic(volume: number) {
+        this.service!.updateCharacteristic(this.volumeCharacteristic!, volume);
     }
 }